implemented (currently disabled) parameters to handle refl redist
these are currently commented out but can easily be added to developer mode as soon as libphoebe accepts any ratios to go to different redistribution methods.

Went for alternative 2 in issue #95.  frac_refl_bol determines the fraction of all light that goes to reflection.  frac_refl_*_bol then determines the fraction of REFLECTION light that goes to each redistribution method.  For now, only noredist is supported (and so is always 1.0).

This closes #95.  A new issue will eventually handle redistribution.