Replacing the use of fortran based atmx and planckint in passbands.py with C/C++ based ones found in libphoebe. I don't know the full functionality passbands.py so I can not test it. Somebody (@kecnrty, @aprsa) needs to check this and correct if there are some mistakes. This should be considered as a draft. Connected to the issue #51.
I also switched to new interp, which is a drop-in replacement of that in /algorithms.
An attempt to address the issue #52

The interface to the new atmx and planckint is a bit different than the ones from fortran.
Only logarithmic intensities are calculated and it accepts numpy arrays of temperatures,
...

Switched off CI-travis, as I don't know if the code works.