fixed broken plot legend labels
closes #106

problem wasn't that the labels were being truncated, but rather that the logic which compared the twigs for the x and y parameter was causing overlap in the end of the parameter.  For example, 'fluxes' and 'times' would still result in keeping the 'es@' at the beginning of the legend label.  This is now fixed.