Revert "added warning (via logger) when accessing param with delayed constraint"
This reverts commit 942eaaf5a3f6e70c0a7d9686b82ca8ee99cce040.

For some reason this commit was causing travis on osx to timeout and error.  I'll try to look into why and re-implement this in the future, but for now will just revert so that builds can be tested again.

related to #154

For some reason this was causing travis on osx to timeout.  I'll try to look into why and re-implement this once I get it fixed