update hierarchy on change_component
related to #20

this doesn't close the issue yet: this will still fail when one component name is a subset of another (ie. starA and starAB) since it does a simple string replace.  Eventually we will need to replace the whole hierarchy framework with something similar to constraints - where the components are actually stored as unique strings and parsed on the fly.