cleanup call to reflection
bolometric intensities and ldints are no longer stored in the mesh, but rather computed on the fly and used immediately.  All reflection-related computations (including computing bolometric intensities) are now all done within handle_reflection which is simply called by populate_observables if necessary.

This could possibly still be optimized for circular cases, where the normal bolometric intensities are not changing in time.... but that would require storing in the mesh again.