rename irradiation effects to match redistribution and handle general case of reflection (#181)
* cleanup call to reflection

bolometric intensities and ldints are no longer stored in the mesh, but rather computed on the fly and used immediately.  All reflection-related computations (including computing bolometric intensities) are now all done within handle_reflection which is simply called by populate_observables if necessary.

This could possibly still be optimized for circular cases, where the normal bolometric intensities are not changing in time.... but that would require storing in the mesh again.

* change names for irradiation to match changes in redistribution branch

frac_refl_bol -> irrad_frac_refl_bol
reflection_method -> irradiation_method
etc.

These are changed now so that when redistribution is supported we won't need to change any syntax.

* implemented general (non-convex) case of reflection into universe

* fixed bug in getting pot for io wrapper

now explicitly requests context='component'.  Without this, if a model with a mesh was previously computed, there was a conflict with the context='model'.

* nosetest for reflection in contact binary

* fixed nosetest to work with issue #110

see #110
contact binaries via the legacy wrapper still require the morphology to be set correctly (they don't listen to they bundle hierarchy as they should).