tweaked behavior of remove_dataset when dataset not passed
for example:
b.remove_dataset(kind='rv') used to leave "phantom" parameters in the compute options that belonged to RV datasets but had kind='phoebe'.  Now these parameters are removed as well.

Current logic:
if dataset is not passed directly, then the filter is applied to determine the matching dataset tags.  All parameters with this dataset tag in the applicable contexts are then removed.