fix issues when times array was out of order, including fti overlap
if fti was enabled by the exptime was longer than the spacing between points, then the oversampling caused the times to no longer be sorted.  This is now fixed and the returned synthetic model is ALWAYS in sorted time order, even if the input data is not. 

When its time to do fitting, we may need to reconsider this as it could cause issues with comparing residuals, etc.