replace pbspan with ptfarea
Definition of flux and luminosity now use ptfarea instead of pbspan.  In the bolometric case, these give the same quantity.  This discrepancy was absorbed entirely by pblum scaling, so relative fluxes should not be affected, but the underlying absolute luminosities were incorrect for passbands (non-bolometric).

In addition to under-the-hood changes, the exposed mesh column for 'pbspan' is now removed and replaced with 'ptfarea'.