temporary hack for renaming colon->long
will only be present in 2.0.* releases, once 2.1 is released we can drop the hack and break backwards compatibility in the release notes.

We don't need to handle the exception in loading a bundle from disk since there was also a bug that prevented features being saved to disk.  Instead, we only need to create a constraint for the duplicate parameter so that scripts raise an intelligent error instead of just not finding the Parameter.