split interactive into interactive_constraints and interactive_checks
Proposed solution to #255

interactive_checks sets whether system checks are run with each line (raising a logger warning) or only at run_compute (at which point they'll raise an error).  By default, this is now set off when running a script, but on when running in an interactive console (or with -i flag).

interactive_constraints is still on by default in all cases (see #154)

phoebe.interactive_on() and phoebe.interactive_off() still exist but set the value of both switches, to keep the previous behavior.

There is now also a phoebe.interactive_checks_on/off and phoebe.interactive_constraints_on/off for overriding the default settings.