implement ability for mpirun to be handled within or outside of PHOEBE
proposed solution to #254

this removes the force_serial switch.  Instead the default is to assume the user is handling any parallelization.  If switched to phoebe.mpi_on() in a serial environment, the job will detach a new mpirun instance using and retrieve the results.  If switched to phoebe.mpi_on() within mpirun, PHOEBE will decide how to handle the parallelization (currently by splitting at the per-time level).