recompute local quantities ALWAYS (for now)
note: this will need to be fixed in the future.  The problem is that we're storing the standard BEFORE filling local quantities.  So when we then detect that we don't need to remesh, we access this standard, but there are no local quantities.  We want to store the local quantities and only recompute them if we need to (i.e. the presence of features).  For now, this will get around the issue, but will cost efficiency in the simplest cases (circular, no features)