created parameters and constraints for requiv_min/max
for contact binaries requiv_min and requiv_max are both exposed, for detach/semi-detached only requiv_max is exposed.  The constraint for requiv_max intelligently (kind of) switches between L1 and L2/3 depending on the morphology of the system.

System checks now account for overflow/underflow appropriately based on morphology and the values of these constrained parameters.