Added __str__ and __repr__ methods for the Passband class.
The __str__ method prints out the passband qualifier, version and tables
that the passband provides. The __repr__ method provides the class name
and the qualifier, i.e. <Passband: Johnson:V>.

In addition, class init parameters reference, version and comments are
now stored in the class. However, they are not saved/loaded because we
need to discuss/converge on the versioning solution first.