implement b.compute_pblums
takes the compute label and computes the pblums (according to coupling, etc) at t0 and returns a dictionary with keys component@dataset and values with the compute passband luminosities.  This is simply a convenience function and is called internally by run_compute even if the user calls it manually.  To access per-time pblums, it is still necessary to add a mesh dataset and expose the pblum column(s).

related to #143