warnings on importing bundle in old version and forbid set_hierarchy
set_hierarchy requires rebuilding constraints, which will result in an infinite loop as the new constraints fail to converge.  It may be possible to migrate old constraints to new constraints, but that will be very complicated if any have been flipped, so instead we forbid calling set_hierarchy.  Old bundles imported should still continue to function in SI with the existing hierarchy.