volume beyond critical hotfix (2.1.3 release) (#316)
Fallback on critical volume if over by within 1e-10.  Numerics in constraints and switching between solar and unitless can result in passing a target volume /slightly/ larger than the critical value (usually at the 12th decimal) which then crashes libphoebe.  To get around this, we now compare the target volume to the critical value from libphoebe, and if it is larger but the relative difference is within 1 part in 10^10, then we adopt the true critical value instead.