add 'auto' (now default) as an option for ld_coeffs_source
will use the value of the corresponding 'atm' parameter, fallingback on ck2004 when necessary.  This is now the default, meaning to set manual ld_coeffs, you need to first set ld_func to something other than 'interp' and then ld_coeffs_source to 'none' in order to access the ld_coeffs parameter.