remove ldint from weights for rvs and lps
the computation of flux from pre-integrated quantities does *not* depend on ldint; only luminosity depends on ldint. As we are weighting RVs and LPs by per-element fluxes, ldint should not be there. Note that intensities and abs_intensities are simply scaled wrt one another, so nothing element-variable is happening there.