consider boosting as an extrinsic effect
however... note that boosting only affects fluxes NOT luminosities (as it only affects Imu and not Inorm).  This means that it will not show up in pblum_ext vs pblum or pbflux_ext vs pbflux and will NOT be accounted for when converting from pbflux when pblum_mode='total flux'.