implement PS.get_info()
accepts the attribute (defaulting to description) and any filter arguments.  Returns a dictionary-like object with qualifiers/twigs for keys and the attributes of the parameters as values.  Entries are merged whenever they have the same attribute-value.  The subclassed dictionary object has a representation similar to the Parameter representation to display nicely when printed in the console.