handle incorrect time vs times usage in plot depending on animate
time should be passed if animate=False, and times should be passed if animate=True.  Now there are relevant warning messages and reasonable fallbacks are taken if the wrong kwarg was passed.

Note that 'time' is still used for filtering if passed to plot with animate=True, so there is a case where the user may want to pass BOTH time and times.