auto add figure options on add_dataset if not already exist for kind
may make this an option to add_dataset or even a setting parameter in the future.  This is probably the desired default for the GUI, but perhaps not if not using figures at all.  Or maybe a setting to disable adding all figure options would be useful - the GUI could then enable the setting by default.