choices for figure parameters now obeys disabling plotting via envar
if plotting is disabled we don't want to import autofig or matplotlib.... so we now fallback on a barebone set of choices.  If plotting is enabled, we now pull directly from the cyclers in autofig.  This also means changing from single letter (k,r,g,b) to full color names (black, red, blue, green)


TODO: we should probably implement some way to update the choices for a change in autofig/MPL version or for creating parameters with plotting disabled and then running that same bundle with plotting enabled.  Alternatively, when plotting is disabled we could skip creating parameters.... but then we may want to automatically create them if plotting is ever re-enabled.