implement b.get_distributions_object with framework for combine logic
combine will be passed to all methods with the ability to act on distribution as a list and will determine how to handle multiple distributions attached to the same parameter.  The solver (fitting) options will then have reasonable defaults, but expose option to switch to any supported behavior.  Currently imagining options for 'first', 'and', 'or'... but 'and'/'or' need to be implemented at the npdists version first before enabling across the board.