support optimizing array parameters
* samplers not yet supported
* individual elements referenced with [index] notation within the twig or qualifier

For example:
* b.get_value('ld_coeffs[0]@primary')
* b.get_value('ld_coeffs@primary[0]')
* b.get_value(qualifier='ld_coeffs[0]', component='primary')

This notation currently works ONLY for get_value and fit_parameters within optimizer solvers (but will eventually be expanded to add_distribution... not sure if filter/get_parameter will be updated to ignore the index or return nothing).

Use b.expand_value(qualifier='fit_parameters') to see how the set value will be adopted and which exact elements in which parameters will be optimized.