kepler's third law constraint: rewrite 1+q->q+1 to be distl-friendly
otherwise float + distribution will fail to propagate distributions through constraints, and there isn't anything we can do about that.  If we want to be able to handle those situations in general, we would need to catch those and rewrite them as distl.delta objects as python will never call distl add methods on a float object.