Implemented numerical integration support for jktebop through the parameters fti_method, fti_oversample, and exptime.
The implementation uses twigs since b.get_value() would fail in odd ways:
  - Either fails because returns a parameter set with both PHOEBE and JKTEBOP parameters
  - Or fails because jktebop parameters are not found if "compute" or "kind" is specified

For some reason, when compute is instead specified using twigs, it works without any issue.

Uses a try except statement to check if fti_oversample and exptime exists, instead of an if statement. Should probably be improved.

Checked that in this implementation, fti_oversample accurately does not exist for RV curves, meaning it is done on a dataset-by-dataset basis.