Fixes the spuriously failing "test_rm" test. (#787)
* Simplified CI run
This workflow simplifies the testing matrix by using python 3.7 through 3.11 on [ubuntu,macos]-latest and pip installing all the latest versions of all dependencies. While less comprehensive than before, it arguably still catches all practical issues.

* Fix test_rm numerical instability
This test compared RVs to legacy for two identical stars, including exactly at conjunctions. That means that stars exactly overlapped, which is numerically unstable. The secondary star is now slightly smaller and slightly less massive. That causes RV of the secondary during secondary eclipse to be nan (because we can't see the star), so that timestamp is nan'd for the legacy as well, for comparison purposes. This now fixes the test and makes it numerically stable.