Interpolation rewrite into C (#792)
* Restructuring of passbands continues.
In this commit the LD interpolator is initialized at passband load-time, and only once. Code has been further streamlined.

* Interpolation redesign in C.
This is a massive commit that replaces the python-prototyped interpolation and extrapolation with a C-based counterpart. Currently, only in-grid interpolation is supported, with the rest to follow. All tests pass and the build is stable. The many modifications were required to sort out shape mismatches and memory alignments necessary for the C extension to reuse the memory rather than reallocate it. High-level details will be added to the final commit.

* Added blending_method to other backends.
This is needed for passing appropriate coefficients to alternate backends.

* Resolving the rebasing errors

* Fixes the spuriously failing "test_rm" test. (#787)

* Simplified CI run
This workflow simplifies the testing matrix by using python 3.7 through 3.11 on [ubuntu,macos]-latest and pip installing all the latest versions of all dependencies. While less comprehensive than before, it arguably still catches all practical issues.

* Fix test_rm numerical instability
This test compared RVs to legacy for two identical stars, including exactly at conjunctions. That means that stars exactly overlapped, which is numerically unstable. The secondary star is now slightly smaller and slightly less massive. That causes RV of the secondary during secondary eclipse to be nan (because we can't see the star), so that timestamp is nan'd for the legacy as well, for comparison purposes. This now fixes the test and makes it numerically stable.

* Interpolation rewritten in C.

* Updated default bundles.

* Added docstrings for *blending_method to alternate backends

* Removed *blending_method parameters from alternate backends
We need a more thorough discussion on how to handle this; see issue 207 (https://github.com/orgs/phoebe-project/projects/2/views/11?pane=issue&itemId=43901138) for details.