Interpolation/extrapolation switch to a C module (#826)
* Implements raise_on_nans behavior and adds a pytest.

* Lingering flag fix.
Per-element index-finding function did not always assign the flag variable, which means that a lingering value could be assigned to the next element; fixed. Also, the return_nanmask parameter was temporarily returned so that the current calling structure is retained. This will be changed soon though.

* Implements nearest neighbor extrapolation
The implementation is done in python and it has been tested for a single vertex dimension. This implementation is meant to serve as a benchmark for the C implementation.

* Added object files to gitignore

* First functional cndpolator migration.
This is a major backend rewrite for interpolation and extrapolation. It will be broken down into several PRs. It passes all tests but several important aspects still need to be completed:
* evaluate the purpose/validity of python-level find_indices() and find_hypercubes();
* remove the need to transfer arrays between python and C on subsequent ndpolator calls;
* further optimize/validate extrapolation to nearest and linear;
* write a test unit.

* Optimization for the nearest neighbor lookup.
The original code built a full N-dimensional lookup grid; the optimized version builds a (N-M)-dimensional grid of basic axes, assuming that attached axes always have their function values defined. Other minor fixes were introduced (docs expanded, variable declaration made more mac-friendly).

* A small nearest extrapolation bugfix.
When the query point was on vertex that was undefined, such a query point was not correctly flagged as out of bounds. This has been fixed, along with a few small redesigns and interp->ndpolate name change. Minor code cleanup done as well.

* Top-level document added.

* Major revision with full functionality
This is another, and hopefully last, full revision with complete interpolation and extrapolation functionality. The code is fully documented, doxygen style. Minimal loose ends still remain, none that would adversely affect the review. So if you've not started yet, now's your chance. :)

* Tidying up debugging statements.
All debugging statements are now controlled by each function's respective debug variable. They are now all set to 0 to avoid verbose output. Note that compiler optimization is still turned off and debugging elements are still compiled in (see setup.py). This means that performance should not be gauged using this code.

* Cashing interpolation tables enabled.
When interpolation tables are built, a pointer to it is stored in a python capsule. This capsule can be dereferenced into the ndp_table structure on subsequent runs. This avoids building the interpolation table each time and thus can save quite a bit of time. Debugging symbols are still on, which leads to ~15% speed reduction.

* Passband computation re-enabled.
Passband computation has been broken and it is now fully functional again. That said, it is quite non-optimal and further work should be done to sort it out. For example, we should no longer store both tables and ndp instances -- just the latter suffices and it will free up half the used memory.

Docs and comments have also been updated.

* Updates TMAP to include all four grids

* Update bundle.py to include all four TMAP grids

* Update ndp_types.h

* Fixed computation of ldcoeffs for TMAP

* Addressing Dave's comments + minor aesthetics.

* Removing effective wavelengths from passband init.
Passband effective wavelengths are no longer used; they were needed before model atmospheres. They were passed to the passband's __init__() function. But the __init__() function computes all quantities required to compute the effective wavelength (photon-weighted and energy-weighted passband area) so it makes no sense to require it as an argument.

In addition, h. c, kB constants were removed from the passband attribute list and are used straight from the units module when necessary.

Finally, several minor aesthetic changes were introduced.

* Quantity/value fix for effective wavelength.

* Separated ndpolator to a standalone package.
Ndpolator is now maintained as a standalone package on [github](https://github.com/aprsa/ndpolator). Phoebe now depends on it and installs it from pip at build time. All tests have also been migrated to the standalone package.

* Update to TMAP abbreviations to keep fits standard

TA = TMAP DA
TO = TMAP DO
TS = TMAP sdO
TM = TMAP DAO

* Added TMAP references to bundle.py

* Update passbands.py

Added missing definition of self.atm_axes in load function

* Blackbody extinction was broken; fixed.

* run_checks_compute() bugfix
The run_checks_compute() function included the internal "_default" compute parameter set in checking the ck2004 model atmosphere table existence in the passband files. This caused the checks to fail even if ck2004 was never needed.

* Fixing blackbody extinction
Several fixes in this commit:
* computation of stock bolometric and Johnson V passbands (this time with blackbody extinction included) added to passbands.py's __main__ block;
* blackbody without extinction does not use any ndpolation tables, but blackbody with extinction does. The query points passed to it were model atmosphere and they did not include rv and ebv correctly; fixed;
* blackbody extinction test raised an exception when extinction tables were not present in the stock Johnson V passband. Instead, the test now computes extinction tables on the fly if they're missing;
* finally, blackbody extinction test has been thoroughly expanded to test both backend and frontend functionality.

* Revert a line that was accidentally commented out.

* Improved fix to the run_checks issue (from 2.4.13).

* Bumping required ndpolator version to 1.1.
Version 1.1 fixes several memory leaks and optimizes the runtime (debugging symbols are excluded and compiler optimization (-O2) is turned on).

---------

Co-authored-by: David Jones <Raindogjones@users.noreply.github.com>