Memory optimization (#854)
* All tables are now only stored in ndp structures.
Previous to this change, the tables were stored both in internal passband structures and in the ndpolator instances, which caused needless duplication of memory consumption. All tables are now stored only in ndpolator instances. In addition, a new test is added to test the generation, saving and loading of passband files.

* Significant optimization of the save() passband method.
The method is now substantially condensed and (hopefully) more readable.

* Memory duplication for axes removed.
There are now no passband structures left that utilize memory needlessly.

* Minor optimizations.

* Changed the passband test to more gracefully account for model atm availability.

* Renaming supported_atms to avoid possible confusion.