New framework for model atmospheres (#1007)
* Implements classes for model atmospheres.
This is a major rewrite of the logic that supports model atmosphere definitions. The parent class, ModelAtmosphere, has been thoroughy documented. Several derived classes are implemented that employ the new concept.

* Removes imputing adds node addition
Imputation of stellar atmospheres should be out of scope for phoebe itself. The capability to impute model atmospheres has been migrated to a standalone code that should be used to update model atmosphere files, not phoebe internals.
For missing values that lie within the convex hull, it is sometimes necessary to add nodes to model atmosphere axes (for example, teff=5000K for phoenix). An add_axis_node() method was added to the ModelAtmosphere class that does this.

* Moves model logic to a separate file
* The logic that pertains to model atmospheres is now in phoebe/atmospheres/models.py.
* Model atmosphere instantiation from fits files is moved to a class method from_path().
* Class attributes of ModelAtmosphere and its subclasses now suffice for passband handling.
* A master list of model atmospheres is in _atmtable.
* pb.save() has been updated to use the new ModelAtmosphere classes.

* Updates load() and save() methods.
I'm not too happy with how associated axes are treated and may need to revisit it further. The next step is to update all passband functions to use atm.basic_axes instead of predefined teff/logg/abun arrays. That will likely shed additional light on how to best treat associated axes.

* Version modification for passband versioning purposes.

* Implements new model atmosphere structures

* Removes associated axes from model atmospheres
Providing associated axes to class-level meta makes little sense because it is table-specific. Thus, all associated axes have been removed from model atmosphere definitions.

* Fixes limb extrapolation, load/save
* Dave reported that limb extrapolation sometimes failed; this was due to Inorms being extrapolated to negative values. This is now fixed.
* passband load/save methods fixed to properly account for the removal of associated axes from the model atmosphere class.

* Fixes a bug in exporting legacy blackbody tables.

* Tweaks to pb.save/load and adding TMAP atm_models (#8)

* Added wls_file option to define name of npy wavelength file

* Updates the code to ndpolator 1.2.2
Ndpolator meta tables were switched from tuples to dictionaries; this commit updates the code to conform to the upstream change. Version 1.2.2 is explicitly requested.

* Fixes backward compatibility, other bugs
This is a simplified version of @Raindogjones' PR in addition to the number of other bugfixes.
* model atmospheres for 'extern_planckint' and 'extern_atmx' implemented;
* save() method simplified to export a pre-2.5-compatible version on demand;
* code updated further to the new model atmosphere framework.

* Updates the blackbody test to the new framework.

* Fixes a failure on save for extern_* models.

* Fixes a bug where LD coeffs could be requested from non-LD atmospheres

* Fixes passbands unit test; removes global atm tables.

* Fixes a typo in test_passbands()

* Fixes missing import rules for pre-2.5 passbands.

* Generalizes query point packing.
This causes issues with tests because overloaded query points are not handled yet in passbands.py so expect breakage.

* Photon-weighted ptf_area definition fix (#10)

* Fix to photon-weighted ptf_area

* Fix to energy-weighted ptf_area

* Fixes a bug in blackbody treatment and further adapts the code to the new framework
The bug in Inorm caused log-intensities instead of intensities to be returned; fixed. This also exposed a difference between the original and the new blackbody treatment that needs to be further investigated.

The same difference causes the discrepancy between interpolated and rigorously computed Sun-Earth intensities to creep up a little bit, so the tolerance had to be increased from 0.1% to 1%, at least temporarily.

Several code tweaks had been done to further adapt the code to the new model atmosphere framework.

* Adds more aggressive version parsing.
It turns out that requiring the version to be < 2.4.17.dev+feature-blending does not suffice because 2.4.17 fails that test. Thus, "<" is now changed to "!=", and will be reverted to 2.5.0 once released.

* Removes explicit listing of supported model atmospheres.

* Introduces labeling of query point columns, fixes multiple bugs
Instead of passing query_pts, the framework now relies on query_table that packs query_pts alongside column names. Each function in passbands.py knows what columns it needs to do its thing, so we can pack anything and everything necessary into query_tables, and respective functions will pick out what they need. This allows us to combine different atmospheres and it greatly simplifies the logic of existing functionality.

* Adds new comparison files for LD tests.

* Fixes two extinction bugs
* the first bug caused blackbody extinction to fail because column concatenation was not done correctly;
* the second bug did not recompute blackbody extinction tables when loading pre-2.5 passbands.

* Final fixes!
After hours of sleuthing and @kecnry's help and keen eye, everything's been fixed up. This makes the PR finally viable.

* Updates stock passband files (bolometric and Johnson V).

* Adds the ability to skip adding history entries to passbands.
This is used when we're recomputing blackbody atm response on the fly when importing from older passbands.

* Removes debugging statements.

Co-authored-by: Kyle Conroy <kyleconroy@gmail.com>

* Removes debugging statement.

Co-authored-by: Kyle Conroy <kyleconroy@gmail.com>

* Removes debugging statements.

Co-authored-by: Kyle Conroy <kyleconroy@gmail.com>

---------

Co-authored-by: David Jones <Raindogjones@users.noreply.github.com>
Co-authored-by: Kyle Conroy <kyleconroy@gmail.com>