Implements blending into backend (#1080)
* Adds a blackbody interpolation test.
The purpose of this test is to make sure that we reach target precision (currently 1.2e-5, but subject to change) across the definition range. This version uses ck2004 temperature span to test (3.5-50kK), but the whole range has been tested with finer step and the maximal deviations remain in this temperature range.

* Fixes an incorrect method comment.

* Simplifies interpolation queries by introducing the InterpQuery class.
Before this change we relied on a (cols, pts) tuple to be passed to the interpolators. The class currently has minimal functionality but it will grow to enable the calling functions to offload most of the boilerplate work to this class.

In addition, superfluous imports in bundle.py have been removed.

* Minor tweaks, unit test for InterpQuery added.

* Introduces InterpResult class, modifies calling functions accordingly.
This is a part of the transition to the more pythonic and human-friendly passband backend. All calling sequences have been updated accordingly. Other changes:
* whitespace in slicing added to make linter happy
* function ldint() renamed to interpolate_ldints() to adhere to the rest of the function naming.

* Switching Inorm and Imu to the new infrastructure.
* Inorm renamed to interpolate_inorms.
* Imu renamed to interpolate_imus.
* Calling functions now use InterpResult consistently everywhere.
* InterpResult unit test added.

* Final touches to blending.
* blending in inorms and imus fully implemented.
* auxiliary functions _log10_Inorm_extern_*  and _log10_Imu have been removed; the computation has been streamlined by using classes and now happens in interpolate_inorms().
* interpolate_inorms() has been pruned and tidied.
* interpolate_imus() has been pruned and tidied.
* comments added to aid with logic flow.
* preliminary (copilot-created) tests for inorms and blending added.

* Simplifies superfluous dictionary comprehension.

* update CI to install jktebop 4.4 (#1079)

(cherry picked from commit 21bb87d909a2fc063dc4237ee2b6bf95d735e567)

---------

Co-authored-by: Kyle Conroy <kyleconroy@gmail.com>