Fixes a pblum-mode bug. (#1091)
* Fixes a pblum-mode bug.
The median value for scaling was not computed correctly before -- it used a full array of fluxes instead just of the reference flux. The code was also simplified as I worked through the logic to understand what's going on. Third light parts have also been updated but not yet tested. All tests pass.

* Adds an explicit model arg to compute_l3s().
This commit catches the circumstance when pblum_mode='dataset-scaled' but no model is passed, which caused the returned l3s to be in absolute units. This makes no contextual sense for the user. A warning is also issued when both model and use_pblums is passed -- pblums take precedence and model is ignored. This closes #1085.

Two stale debug print statements from the previous commit have been removed.

A thorough, claude-generated compute_l3s test has been added to the test suite.

* Names arguments to set_value.

Co-authored-by: Kyle Conroy <kyleconroy@gmail.com>

* Names arguments to set_value.

Co-authored-by: Kyle Conroy <kyleconroy@gmail.com>

* Apply suggestions from code review

Co-authored-by: Kyle Conroy <kyleconroy@gmail.com>

* Removes a debugging relic that caused checks to be skipped needlessly.

* Reverts file rename, renames a function instead.
Also makes Kyle happy. ;)

* ChangeLog updates.

---------

Co-authored-by: Kyle Conroy <kyleconroy@gmail.com>